backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain==0.3.15
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-chroma
langchain-core
pypdf
beautifulsoup4
pydantic
python-multipart
